chore: UI code refactoring for better reuse
